requirements.txt
pyyaml>=5.4
requests>=2.26
sqlalchemy>=1.4
goodfaith>=0.1.0
suby>=0.0.3
